+++ processing/crop.go
@@ -53,8 +53,7 @@ func crop(pctx *pipelineContext, img *vips.Image, po *options.ProcessingOptions,
 
 func cropToResult(pctx *pipelineContext, img *vips.Image, po *options.ProcessingOptions, imgdata *imagedata.ImageData) error {
 	// Crop image to the result size
-	resultWidth := imath.Scale(po.Width, po.Dpr)
-	resultHeight := imath.Scale(po.Height, po.Dpr)
+	resultWidth, resultHeight := resultSize(po)
 
 	if po.ResizingType == options.ResizeFillDown {
 		if resultWidth > img.Width() {